Allow select to work with promise
Modified the select method to work with promises while preserving the original functionality of inserting a string that wasn't a promise.

This allows for us to use the `async` prefix on the insert method when defining a custom insert method that gets passed to the options. 

The reason this is necessary is because it enables us to execute an action asynchronously prior to inserting the new element. For example, creating a new tag, then modifying the insert content with the async response data.